Remover SAML do Redash
Ajustes para gerar exe com pyinstaller
