+++ redash/templates/login.html
@@ -20,10 +20,6 @@
       </a>
     {% endif %}
 
-    {% if show_saml_login %}
-      <a href="{{ url_for('saml_auth.sp_initiated', org_slug=org_slug, next=next) }}" class="login-button btn btn-default btn-block">SAML Login</a>
-    {% endif %}
-
     {% if show_remote_user_login %}
       <a href="{{ url_for('remote_user_auth.login', org_slug=org_slug, next=next) }}" class="login-button btn btn-default btn-block">Remote User Login</a>
     {% endif %}
@@ -33,7 +29,7 @@
     {% endif %}
 
     {% if show_password_login %}
-    {% if show_google_openid or show_saml_login or show_remote_user_login or show_ldap_login %}
+    {% if show_google_openid or show_remote_user_login or show_ldap_login %}
     <hr>
     {% endif %}
     <div class="col-sm-6 d-flex flex-column justify-content-center align-items-center" style="border-right: 1px solid #ccc; height: -webkit-fill-available;">
