+++ requirements.txt
@@ -38,7 +38,7 @@ rq==1.1.0
 rq-scheduler==0.9.1
 jsonschema==3.1.1
 RestrictedPython==5.0
-pysaml2==4.8.0
+# pysaml2==4.8.0
 # pycrypto==2.6.1
 pycryptodome==3.9.7
 funcy==1.13
